refactor: consolidate gate hook tests into parameterised modules by concern
Replaces 4 gate-specific test files with 6 concern-organised modules plus
a shared helper module. Channel-routing tests now parameterise across all
stop-gates (handover, qa, ida) instead of only ida. Adding a new gate
requires one entry in STOP_GATES.

Structure:
- gate_helpers.py — shared constants, fixtures, subprocess runners
- test_gate_verdict_logic.py — mode overrides × gates, read-only bypass
- test_gate_lifecycle.py — per-gate state transitions (IDA, handover)
- test_output_channel_routing.py — advisory routing × gates × verdicts
- test_schema_conformance.py — platform schema validation × events
- test_subprocess_format.py — JSON format per platform (subprocess)
- test_silent_events.py — task-notification transparency

Closes aops-339da1dc
